pyproject.toml
[project]
name = "ood-slam"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch==2.4.1",
    "hydra-core>=1.3.2",
    "hydra-submitit-launcher>=1.2.0",
    "wandb>=0.17.6",
    "lightning>=2.4.0",
    "torchvision>=0.19.0",
    "rich>=13.7.1",
    "tqdm>=4.66.5",
    "hydra-zen==0.13.1rc1",
    "matplotlib>=3.9.2",
    "optree>=0.15.0",
    # Hugging Face dependencies:
    "evaluate>=0.4.2",
    "transformers>=4.44.0",
    "datasets>=2.21.0",
    # Hydra / submitit plugins
    "hydra-colorlog>=1.2.0",
    "remote-slurm-executor",
    "hydra-auto-schema>=0.0.7",
    "hydra-orion-sweeper>=1.6.4 ; python_full_version < '3.11'",
]

[dependency-groups]
dev = [
    "pytest>=8.3.2",
    "ruff>=0.6.0",
    "pre-commit>=3.0.0",
    "pytest-cov>=5.0.0",
    "pytest-xdist>=3.6.1",
]

[tool.pytest.ini_options]
testpaths = ["ood_slam", "tests"]

# Required to use torch deterministic mode
env = ["CUBLAS_WORKSPACE_CONFIG=:4096:8"]
addopts = [
    "--doctest-modules",
]

[tool.ruff]
line-length = 99
indent-width = 4
lint.select = ["E4", "E7", "E9", "F", "I", "UP"]

[tool.uv]
managed = true
python-preference = "system"

## From https://docs.astral.sh/uv/reference/settings/#index-strategy:
## "Only use results from the first index that returns a match for a given package name."
## In other words: only get the package from PyPI if there isn't a version of it in the DRAC wheelhouse.
# index-strategy = "first-index"

## "Search for every package name across all indexes, exhausting the versions from the first index before
##  moving on to the next"
## In other words: Only get the package from PyPI if the requested version is higher than the version
## in the DRAC wheelhouse.
# index-strategy = "unsafe-first-match"

## "Search for every package name across all indexes, preferring the "best" version found.
##  If a package version is in multiple indexes, only look at the entry for the first index."
## In other words: Consider all versions of the package DRAC + PyPI, and use the version that best matches
## the requested version. In a tie, choose the DRAC wheel.
index-strategy = "unsafe-best-match"

[[tool.uv.index]]
name = "drac-gentoo2023-x86-64-v3"
url = "/cvmfs/soft.computecanada.ca/custom/python/wheelhouse/gentoo2023/x86-64-v3"
format = "flat"

[[tool.uv.index]]
name = "drac-gentoo2023-generic"
url = "/cvmfs/soft.computecanada.ca/custom/python/wheelhouse/gentoo2023/generic"
format = "flat"

[[tool.uv.index]]
name = "drac-generic"
url = "/cvmfs/soft.computecanada.ca/custom/python/wheelhouse/generic"
format = "flat"

[tool.uv.sources]
remote-slurm-executor = { git = "https://github.com/adamburhan/remote-slurm-executor", branch = "master" }
